hwui: Vulkan: Destroy semaphores after retrieving fd
Debuging shows that flush() destroys semaphores before
mGetSemaphoreFdKHR is called.
In turn, the following errors show during normal usage:

    OpenGLRenderer: VulkanManager::swapBuffers(): Failed to get semaphore Fd

Using a camera application however results in extremely sluggish frame
updates (~2 fps) with the following error constantly printed:
OpenGLRenderer: VulkanManager::createReleaseFence: Failed to get
semaphore Fd

To resolve this, d92a9b158e8a473cd7860f2ea6cc9090fc294f78 can be
reverted which postpones destroying semaphores until after an FD has
been created, or the current code altered to postpone this as has been
done in this commit.
Furthermore the deleted comments in that commit clarify that it is safe
to remove the semaphore as ownership was transfered to the Fd, while the
underlying handle still lives on for proper completion signaling when
enqueued on a command buffer.

Fixes: d92a9b158e8a473cd7860f2ea6cc9090fc294f78
Change-Id: I750a26cd31e1802be70840fa86e47f1bc4e91942
Signed-off-by: Marijn Suijten <marijns95@gmail.com>